+++ kernels/neuroFunctions.cu
@@ -1,23 +1,23 @@
 #include <neuroFunc.h>
 #include <utilFunc.h>
 
-__host__ __device__ void neuroSum(float &store, float &input){
+__host__ __device__ void neuroSum(double &store, double &input){
     store += input;
 }
 
-__host__ __device__ void neuroZero(float &store){
+__host__ __device__ void neuroZero(double &store){
     store =0;
 }
 
-__host__ __device__ void neuroSquash(float &store){
+__host__ __device__ void neuroSquash(double &store){
     store = ActFunc(store);
 }
 
-__host__ __device__ void neuroMemGate(float &memIn, float &input, float &output, float min){
+__host__ __device__ void neuroMemGate(double &memIn, double &input, double &output, double min){
     if(memIn > min || memIn < -min)
         neuroSum(output, input);
 }
-__host__ __device__ void neuroMemForget(float &memForget, float &mem, float min){
+__host__ __device__ void neuroMemForget(double &memForget, double &mem, double min){
     if(memForget > min || memForget < -min)
         neuroZero(mem);
